Memory Efficient Optimizations, Double Down Logic Continued, Bug Fixes, & if condition reorganization
- reorganized game conditions statements, some werent checked in the correct order
- double down flag toggle for visual change on its card. When player double downs the card will be side ways (implemented soon)
- fixed a bug where double down button was pointing to the onHit listener instead of the onDoubleDown listener
 - small type changes in for loops for displaying cards
- Type changes from int to uint8_t for memory efficiency, aproxx. 75% memory reduction for variables that go from type int to uint8_t

+++ src/card.cpp
@@ -11,7 +11,7 @@ const std::string Card::displayCmd()
     return getSuit() + " " + getRank();
 }
 
-int Card::getValue() const // for adding up cards
+uint8_t Card::getValue() const // for adding up cards
 {
     switch (rank)
     {
